A working tile selector.
This is a tile selector program where it can create X amount of tile(s) and adjusting its layout according to the programmers' wishes. In this case, it is set to a 3x5 tile with 15 tiles within it. When run, calling (Test 15), the user are able to use the arrow keys to select the tiles. Note: 15 could be change in (Test 15) into any number between 1-15.