Second Mile Stone
This was mainly focusing on creating the different objects and implementing the main rules.
The new objects that were added are player and option. Also, with the new implementation, you can only place tiles that are not already your own color and also adjacent to your tile. For blue team, you start placing tiles from the top castle, and as for the red team, you start placing tiles from the bottom castle. The player object is updated everytime a key is press and checks the board for the number of tiles each player owns. Each player also gain a gold for every new tile conquered. Currently, the implementation of cost/purchasing minions has not been added yet as it is a work in progress.