Castle Siege Alpha
This is the first look and the first version of the working game.
CONTROLS:
-------------------------
During option/menu selection:
Up arrow = Moves the choice selection up, unless it is on the highest one already
Down arrow = Moves the choice selection down, unless it is on the lowest one already
Enter = Chooses that particular choice in the menu *Note: You cannot choose an option that you are not eligible to choose. ex. trying to buy a cost 3 minion when having only 1 gold.*
After choosing cost 1 minions:
Left arrow = Moves the selection to the left, unless it is on the left most already
Right arrow  = Moves the selection to the right, unless it is on the right most already
Enter = create/spawn a minion that has the ability to move once per turn except for the spawning turn at that particular tile selected
After choosing cost 3 minions:
Left arrow = Moves the selection to the left, unless it is on the left most already
Right arrow  = Moves the selection to the right, unless